% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{api_key_check}
\alias{api_key_check}
\title{A check that the API key is set
Called in other functions as an initial check}
\usage{
api_key_check()
}
\description{
A check that the API key is set
Called in other functions as an initial check
}
\examples{
api_key_check()
}
