% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{application_practices}
\alias{application_practices}
\title{Return the list of practices this app is installed on}
\usage{
application_practices()
}
\value{
A list of ints of the practices
}
\description{
Return the list of practices this app is installed on
}
\examples{
application_practices()
}
