% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{create_patient}
\alias{create_patient}
\title{Create a new patient. The patient will be associated
with the provider corresponding to the Thrive API key
in current use}
\usage{
create_patient(first_name, last_name, practice_id)
}
\arguments{
\item{first_name}{First Name of patient}

\item{last_name}{Last Name of patient}
}
\value{
An object with new patient id
}
\description{
Create a new patient. The patient will be associated
with the provider corresponding to the Thrive API key
in current use
}
\examples{
create_patient("Foo","Bar",1)
}
