% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{delete_file}
\alias{delete_file}
\title{Delete a file associated with a patient. Use caution when using this function
as it should only be called to correct mistaken uploads}
\usage{
delete_file(patient_id, uuid)
}
\arguments{
\item{uuid}{String representing the UUID of the file to delete}

\item{patent_id}{patient_id for whom the file is being deleted}
}
\value{
The JSON response
}
\description{
Delete a file associated with a patient. Use caution when using this function
as it should only be called to correct mistaken uploads
}
\examples{
delete_patient_file(599,"d4aa531a-89a0-4720-9060-bb32fa6cecec")
}
