% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_patient_file_list}
\alias{get_patient_file_list}
\title{Get all of the files associated with a patient}
\usage{
get_patient_file_list(patient_id)
}
\arguments{
\item{patient_id}{the id of the patient to obtain the list for}
}
\value{
graphql_query S3 object
}
\description{
Get all of the files associated with a patient
}
\examples{
get_patient_file_list(4)
}
