% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_patient_ids}
\alias{get_patient_ids}
\title{Get all of the patient ids for a practice}
\usage{
get_patient_ids(practice_id)
}
\arguments{
\item{practice_id}{The practice_id to obtain the list}
}
\value{
graphql_query S3 object
}
\description{
Get all of the patient ids for a practice
}
\examples{
get_patient_ids(1)
}
