% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_patient_reports}
\alias{get_patient_reports}
\title{Get all of the patient reports for a practice.}
\usage{
get_patient_reports(project_id)
}
\arguments{
\item{practice_id}{The practice id to obtain reports for}
}
\value{
graphql_query S3 object
}
\description{
Get all of the patient reports for a practice.
}
\examples{
get_patient_reports(1)
}
