% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{graphql_query}
\alias{graphql_query}
\title{Run a GraphQL Query on the web api}
\usage{
graphql_query(query)
}
\arguments{
\item{query}{A properly formatted GraphQL query}
}
\value{
A R object representing the JSON response
}
\description{
Run a GraphQL Query on the web api
}
