% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{post_file}
\alias{post_file}
\title{Upload a file to a patient.}
\usage{
post_file(patient_id, filename)
}
\arguments{
\item{patient_id}{patient_id to upload file to}

\item{filename}{local file to upload}
}
\description{
Upload a file to a patient.
}
\examples{
post_patient-file(599,"new_report.pdf")
}
