% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thriver_package.R
\docType{package}
\name{thriver-package}
\alias{thriver}
\alias{thriver-package}
\title{thriver: thriver}
\description{
To learn more about thriver, view the vignettes
`browseVignettes(package = "thriver")`
}
\author{
\strong{Maintainer}: James Borden \email{james@insilica.co}

}
