% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{get_file}
\alias{get_file}
\title{Get a patient file from the server and
save it locally in getwd().}
\usage{
get_file(url, filename)
}
\arguments{
\item{url}{A url representing the location of the file on the Thrive server
Use `get_patient_file_list` to get the proper url and filename for
files associated with a patient}

\item{filename}{filename to save data to locally}
}
\description{
Get a patient file from the server and
save it locally in getwd().
}
\examples{
get_patient_file("/files/patient/599/d4aa531a-89a0-4720-9060-bb32fa6cecec","report.pdf")
}
